+++ kash_macros/src/io_kash/cache_fn.rs
@@ -61,8 +61,8 @@ impl ToTokens for CacheFn<'_> {
         );
         let cache_name = cache_ident.to_string();
 
+        let set_cache_block = gen_set_cache_block(&self.args.disk, asyncness);
         let return_cache_block = gen_return_cache_block();
-        let set_cache_block = gen_set_cache_block(self.args.disk, asyncness);
 
         let cache_create = gen_cache_create(self.args, asyncness, &cache_ident, cache_name);
 
@@ -71,8 +71,8 @@ impl ToTokens for CacheFn<'_> {
         } else {
             quote! {}
         };
-        let use_trait = gen_use_trait(asyncness, self.args.disk);
-        let async_cache_get_return = if asyncness.is_some() && !self.args.disk {
+        let use_trait = gen_use_trait(asyncness, &self.args.disk);
+        let async_cache_get_return = if asyncness.is_some() && self.args.disk.is_none() {
             quote! {
                 if let Some(result) = cache.get(&key).await? {
                     #return_cache_block